aider: Refactored "ydl_opts" in `utils.py` to be defined in a separate file called "options.py" and modified the code accordingly.
# Context:
USER: refactor the "ydl_opts" used in `utils.py` so it is not repeated in the code. We can extend or modify the options within each method where it is used but the definition should be just one. Create a file called "options.py" or "settings.py" where "ydl_opts" variable is defined. And any others if necessary. Modify the code accordingly.